pyproject.toml
[project]
name = "finrobot-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "finrobot>=0.1.5",
    "google-generativeai>=0.8.6",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "yfinance>=1.1.0",
]
